+++ CHANGELOG.md
@@ -7,11 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed:
+
+- fixed a bug where `log_prefix` can only be a string [#393](https://github.com/Draegerwerk/sdc11073/issues/393)
+
+## [2.1.0] - 2024-09-04
+
 ### Added
 
 - support for python version 3.12
 - new method ContextStateTransaction.disaccociate_all
 - additional schemata for validation can be declared in SdcProviderComponents and SdcConsumerComponents
+- new method Consumer.restart
+- log a message with level INFO when a context state update is received
+- allow to use full qualified name instead of IP in provider and consumer
 
 ### Fixed
 
@@ -30,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - incorrect BindingMdibVersion and UnbindingMdibVersion [#168](https://github.com/Draegerwerk/sdc11073/issues/168)
 - ensure_location_context_descriptor and ensure_patient_context_descriptor also work for multiple system contexts in mdib.
 - provider mdib observables are updated [#365](https://github.com/Draegerwerk/sdc11073/issues/365)
+- fixed possible error if _ThreadingHTTPServer fails to start and in server_close self.dispatcher member does not exist
 
 ### Changed
 
@@ -38,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SetContextState operation sets ContextAssociation according to value in proposed context state. 
   Before the proposed state was always associated. 
   Check added in SetContextState that max. one proposed state per descriptor is associated.
-
+- refactored ClientMdib.reload_all and processing of notifications
+- 
 ## [2.0.1] - 2024-02-21
 
 ### Fixed